+++ history.md
@@ -1,6 +1,8 @@
 ### History
 
 v2.8.0
+- Retain time series across active layer changes.
+- Add time series hover readout.
 - Added marker type and stroke te settings.
 - Previewed polygon while drawing.
 - Added support for double-click to finish a polygon.